+++ src/borgitory/services/notifications/types.py
@@ -5,7 +5,7 @@ Type definitions for notification system.
 from dataclasses import dataclass
 from typing import Dict, Optional
 from enum import Enum
-from borgitory.types import ConfigDict
+from borgitory.custom_types import ConfigDict
 
 
 class NotificationType(str, Enum):
